+++ backend/src/main/java/com/campito/backend/dto/CuotaResumenDTO.java
@@ -1,5 +1,7 @@
 package com.campito.backend.dto;
 
+import java.math.BigDecimal;
+
 /**
  * DTO simplificado para las cuotas dentro de un resumen.
  * Contiene solo la información relevante para mostrar en el detalle del resumen.
@@ -7,7 +9,7 @@ package com.campito.backend.dto;
 public record CuotaResumenDTO(
     Long id,
     Integer numeroCuota,
-    Float montoCuota,
+    BigDecimal montoCuota,
     String descripcion,
     Integer totalCuotas,
     String motivo
