+++ backend/src/main/java/com/campito/backend/dto/EspacioTrabajoDTOResponse.java
@@ -1,11 +1,12 @@
 package com.campito.backend.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record EspacioTrabajoDTOResponse(
     UUID id,
     String nombre,
-    Float saldo,
+    BigDecimal saldo,
     UUID usuarioAdminId
 ) {
